Move unit tests under correct package
