+++ tests/unit/services/dynamodb-images.unitTest.ts
@@ -1,10 +1,10 @@
-import {DynamoDbImage} from "../../src/services/dynamodb-images";
-import {default as primitivesJson} from "../resources/dynamodb-image-primitives.json";
-import {default as setsJson} from "../resources/dynamodb-image-sets.json";
-import {default as listJson} from "../resources/dynamodb-image-list.json";
-import {default as mapJson} from "../resources/dynamodb-image-map.json";
-import {default as nestedJson} from "../resources/dynamodb-image-nested.json";
-import {castToImageShape} from "../utils";
+import {DynamoDbImage} from "../../../src/services/dynamodb-images";
+import {default as primitivesJson} from "../../resources/dynamodb-image-primitives.json";
+import {default as setsJson} from "../../resources/dynamodb-image-sets.json";
+import {default as listJson} from "../../resources/dynamodb-image-list.json";
+import {default as mapJson} from "../../resources/dynamodb-image-map.json";
+import {default as nestedJson} from "../../resources/dynamodb-image-nested.json";
+import {castToImageShape} from "../../utils";
 
 describe("parse()", () => {
     it("should parse a Dynamo image containing all primitive types", () => {
